requirement.txt
requests
pdfplumber
python-dotenv
pymongo